syscalls (#7)
* Ensure pmm_alloc returns physical addresses, not higherhalf ones

* Fix opcode detection in #UD handler

* buddy: fix indentation

* misc: Fix indentation

* vmm: Remove PAT & cleanup pagefault handler

* gnu_no_return attribute:  __attribute__((no_return)) void -> __attribute__((no_return))

* Rewrite vmm, fix some bugs, add some syscall code, general cleanup

* Forgot about this comment, lol.

* Update vmm.c

* vmm: Cleanup

* uspace: I forgot to jump to userspace, oops :^)

* all: vscode didn't save the files automatically, oops

* paging: Remove PAT for now

* Don't swapgs here

* boot: Make fterm great again :^)

* Cleanup folder structure

* Set stack to 123h for testing

* paging: Improved Pagefault handler

* misc: Forgot to remove this comment

* gdt: Forgot a field in USER_DS_64

* int: Actually handle pagefaults now

* vmm: Introduce fast mapping

* daemon_load: swapgs before jumping to ring3

* elf: Simplfy elf loading/mapping

* ustack is undefined(0) until the ELF file has been mapped

* libs/userspace -> userspace/

* uspace: Fix ustack setup

* kernel_syscalls: Remove MSR checks (they always exist when in 64 bit mode)

* kernel_syscalls: Setup user segment selectors before returning to ring3

* cpu: Seperate the iframe struct from the gp_registers_t struct

* Use a better font for the terminal

* percpu: Add a real syscall stack

* serial: Fix comparison of different sized types

* git: Remove .vscode folder

* Refactor code and implement a syscall interface

* Update README

+++ kernel/devices/serial/serial.c
@@ -37,7 +37,7 @@ void serial_write(char chr)
 
 void serial_write_str(const char *str)
 {
-    for (int i = 0; i < strlen(str); i++)
+    for (uint64_t i = 0; i < strlen(str); i++)
     {
         serial_write(str[i]);
     }
